Dilemma refactoring (#259)
* Converted Archer to JSON definition. Included some refactoring of the OvercomeDilemmaConditionAction, which may have fixed an un-noticed bug. Added unit test for Archer.

* Added Parem (Premiere), Parem (TNG), and Karina cards. Added unit test for Parem (TNG).

* Refactored ValueSource class hierarchy to use an interface at the top

* Removed evaluateExpression method from ValueSource classes that represent a range, so we can stop throwing errors

* Removed error throws from ValueSource class

* Added unit test for Picking Up the Pieces. Fixed bugs in presence calculation.

* Removed Java definition for Picking Up the Pieces. Removed NIU class MemoryIdFilter.

* Moved Wind Dancer definition to JSON and added unit test

* Moved Blended to JSON definition

* Moved New Essentialists to JSON definition

* Added partial implementation of full Friendly Fire effect. Still need to work on countdown.

* Added a catch if duplicate blueprint ids are added to the library

* Moved Pinned Down to JSON definition and added unit test

* Added unit test for Pinned Down. Fixed some bugs related to random selections of multiple cards.

* Pinned Down test in good shape

* Converted Homefront II set to raw JSON

* Added unit test for Dangerous Climb

* Moved Tense Negotiations to JSON definition. Added unit test and underlying functionality for nullifying dilemmas.

* Moved Microbrain to JSON definition

* Removed FailDilemmaAction class no longer in use

* Moved Dignitaries and Witnesses to JSON definition

+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/actions/choose/SelectNumberAction.java
@@ -18,8 +18,8 @@ public class SelectNumberAction extends MakeDecisionAction {
     public SelectNumberAction(DefaultGame cardGame, ActionContext context, String choiceText, ValueSource valueSource,
                               String memoryId) throws InvalidGameLogicException {
         super(cardGame, context.getPerformingPlayerId(), choiceText);
-        _minimum = (int) valueSource.getMinimum(cardGame, context);
-        _maximum = (int) valueSource.getMaximum(cardGame, context);
+        _minimum = valueSource.getMinimum(cardGame, context);
+        _maximum = valueSource.getMaximum(cardGame, context);
         _memoryId = memoryId;
     }
